Refactor indicators module and update imports
